added opencl cpp bindings as a dependency rather than storing them locally

+++ common/include/common/frequency_domain_filter.h
@@ -58,14 +58,14 @@ private:
 
 /// See antoni2010 equations 19 and 20
 
-float lower_band_edge(float centre, float p, float P, size_t l);
-float upper_band_edge(float centre, float p, float P, size_t l);
+double lower_band_edge(double centre, double p, double P, size_t l);
+double upper_band_edge(double centre, double p, double P, size_t l);
 
-float band_edge_frequency(int band, size_t bands, float lower, float upper);
+double band_edge_frequency(int band, size_t bands, double lower, double upper);
 
 template <size_t bands>
-std::array<float, bands + 1> band_edge_frequencies(float lower, float upper) {
-    std::array<float, bands + 1> ret;
+std::array<double, bands + 1> band_edge_frequencies(double lower, double upper) {
+    std::array<double, bands + 1> ret;
     for (auto i{0u}; i != bands + 1; ++i) {
         ret[i] = band_edge_frequency(i, bands, lower, upper);
     }
@@ -73,8 +73,8 @@ std::array<float, bands + 1> band_edge_frequencies(float lower, float upper) {
 }
 
 template <size_t bands>
-std::array<float, bands + 1> band_centre_frequencies(float lower, float upper) {
-    std::array<float, bands + 1> ret;
+std::array<double, bands + 1> band_centre_frequencies(double lower, double upper) {
+    std::array<double, bands + 1> ret;
     for (auto i{0ul}; i != ret.size(); ++i) {
         ret[i] = band_edge_frequency(i * 2 + 1, bands * 2, lower, upper);
     }
@@ -82,10 +82,10 @@ std::array<float, bands + 1> band_centre_frequencies(float lower, float upper) {
 }
 
 template <size_t bands>
-std::array<float, bands + 1> band_edge_widths(float lower,
-                                              float upper,
-                                              float overlap) {
-    std::array<float, bands + 1> ret;
+std::array<double, bands + 1> band_edge_widths(double lower,
+                                              double upper,
+                                              double overlap) {
+    std::array<double, bands + 1> ret;
     for (int i{0}; i != ret.size(); ++i) {
         ret[i] = std::abs(
                 (band_edge_frequency(i * 2, bands * 2, lower, upper) -
@@ -101,19 +101,19 @@ std::array<float, bands + 1> band_edge_widths(float lower,
 /// upper: the normalized upper band edge frequency of this band
 /// upper_edge_width: half the absolute width of the upper crossover
 /// l: the slope (0 is shallow, higher is steeper)
-float compute_bandpass_magnitude(float frequency,
-                                 float lower,
-                                 float lower_edge_width,
-                                 float upper,
-                                 float upper_edge_width,
+double compute_bandpass_magnitude(double frequency,
+                                 double lower,
+                                 double lower_edge_width,
+                                 double upper,
+                                 double upper_edge_width,
                                  size_t l);
 
-float compute_lopass_magnitude(float frequency,
-                               float cutoff,
-                               float width,
+double compute_lopass_magnitude(double frequency,
+                               double cutoff,
+                               double width,
                                size_t l);
 
-float compute_hipass_magnitude(float frequency,
-                               float cutoff,
-                               float width,
+double compute_hipass_magnitude(double frequency,
+                               double cutoff,
+                               double width,
                                size_t l);
